userbundle: - utilisation de la class Model\User pour identifier l'entité user, permet ainsi d'étendre du model de base et de profiter des fonctionalités de BigfootUserBundle - déplacement des roles au niveaux de l'entité et non plus la superclass, ceci permet de pouvoir instancier une autre class heritant du model user (avant duplication de la table bigfoot_user_role, car la relation etait au niveau du modèle)
